update to fix increased size of data

+++ Array_Maker.cpp
@@ -8,30 +8,30 @@ using namespace std;
 Array_Maker::Array_Maker() {}
 Array_Maker::~Array_Maker() {}
 
-void Array_Maker::randArray(vector<int>& arr, int n) 
+void Array_Maker::randArray(vector<int>& arr, long long n) 
 {
 	arr.resize(n);
-	for (int i = 0; i < n; i++) arr.at(i) = rand() % 1000;
+	for (long long i = 0; i < n; i++) arr.at(i) = rand() % 1000;
 
 }
 
-void Array_Maker::halfsortedArray(vector<int>& arr, int n) 
+void Array_Maker::halfsortedArray(vector<int>& arr, long long n) 
 {
 	arr.resize(n);
-	int mid = n / 2;
+	long long mid = n / 2;
 
-	for (int i = 0; i < mid; i++) arr.at(i) = i;
+	for (long long i = 0; i < mid; i++) arr.at(i) = i;
 
 	arr.at(mid) = mid;
 
-	for (int i = mid + 1; i < n; i++) arr.at(i) = (rand() % 1000) + arr.at(i - 1);
+	for (long long i = mid + 1; i < n; i++) arr.at(i) = (rand() % 1000) + arr.at(i - 1);
 
 }
 
-void Array_Maker::sortedArray(vector<int>& arr, int n) 
+void Array_Maker::sortedArray(vector<int>& arr, long long n) 
 {
 	arr.resize(n);
-	for (int i = 0; i < n; i++) arr.at(i) = i;
+	for (long long i = 0; i < n; i++) arr.at(i) = i;
 
 }
